MDEV-5536: add systemd socket activation
Systemd has a socket activation feature where a mariadb.socket
defination defines the sockets to listen to, and passes those
file descriptors directly to mariadbd to use when a connection
occurs.

The new functionality is utilized when starting as follows:

  systemctl start mariadb.socket

The mariadb.socket defination only needs to contain the network
information, ListenStream= directives, the mariadb.service
defination is still used for service instigation.

When mariadbd is started in this way, the socket, port, bind-address
backlog are all assumed to be self contained in the mariadb.socket
defination and as such the mariadb settings and command line
arguments of these network settings are ignored.
See man systemd.socket for how to limit this to specific ports.

Extra ports, those specified with extra_port in socket activation
mode, are those with a FileDescriptorName=extra. These need
to be in a separate service name like mariadb-extra.socket and
these require a Service={mariadb.service} directive to map to the
original service. Extra ports need systemd v227 or greater
(not RHEL/Centos7 - v219) when FileDescriptorName= was added,
otherwise the extra ports are treated like ordinary ports.

The number of sockets isn't limited when using systemd socket activation
(except by operating system limits on file descriptors and a minimal
amount of memory used per file descriptor). The systemd sockets passed
can include any ownership or permissions, including those the
mariadbd process wouldn't normally have the permission to create.

This implementation is compatible with mariadb.service definations.
Those services started with:

  systemctl start mariadb.service

does actually start the mariadb.service and used all the my.cnf
settings of sockets and ports like it previously did.

+++ sql/mysqld.cc
@@ -1351,6 +1351,7 @@ struct my_rnd_struct sql_rand; ///< used by sql_class.cc:THD::THD()
 
 Dynamic_array<MYSQL_SOCKET> listen_sockets(PSI_INSTRUMENT_MEM, 0);
 bool unix_sock_is_online= false;
+static int systemd_sock_activation; /* systemd socket activation */
 /**
   Error reporter that buffer log messages.
   @param level          log message level
@@ -1697,8 +1698,10 @@ static void close_connections(void)
   {
     MYSQL_SOCKET *sock= listen_sockets.get_pos(i);
     (void) mysql_socket_close(*sock);
-    if (sock->is_unix_domain_socket)
+    if (sock->is_unix_domain_socket && !systemd_sock_activation)
+    {
       (void) unlink(mysqld_unix_port);
+    }
   }
   listen_sockets.free_memory();
   mysql_mutex_unlock(&LOCK_start_thread);
@@ -2385,6 +2388,62 @@ static void activate_tcp_port(uint port,
   DBUG_VOID_RETURN;
 }
 
+/**
+   Activate usage of a systemd activated sockets
+   i.e started by mariadb.socket
+*/
+
+static void use_systemd_activated_sockets()
+{
+  MYSQL_SOCKET sock;
+  int fd;
+  char **names = NULL;
+  int sd_sockets;
+  DBUG_ENTER("use_systemd_activated_sockets");
+
+  sd_sockets= sd_listen_fds_with_names(0, &names);
+
+  if (!sd_sockets)
+  {
+    DBUG_VOID_RETURN;
+  }
+  DBUG_PRINT("general",("Systemd listen_fds is %d", sd_sockets));
+  while (sd_sockets--)
+  {
+    fd= SD_LISTEN_FDS_START + sd_sockets;
+    if (sd_is_socket_unix(fd,  SOCK_STREAM, 1, NULL, 0))
+    {
+      sock= mysql_socket_fd(key_socket_unix, fd);
+      sock.is_unix_domain_socket= 1;
+    }
+    else if (sd_is_socket_inet(fd, AF_UNSPEC, SOCK_STREAM, 1, 0))
+    {
+      sock= mysql_socket_fd(key_socket_tcpip, fd);
+      sock.is_unix_domain_socket= 0;
+    }
+    else
+    {
+      sql_print_error("Unknown systemd socket activation socket %d,"
+                      " not listening, not unix or TCP socket, or not a type SOCK_STREAM", fd);
+      if (names)
+      {
+        free(names);
+      }
+      unireg_abort(1);
+    }
+    sock.is_extra_port= names && strcmp(names[sd_sockets], "extra") == 0;
+    mysql_socket_set_thread_owner(sock);
+    listen_sockets.push(sock);
+  }
+  systemd_sock_activation= 1;
+  if (names)
+  {
+    free(names);
+  }
+
+  DBUG_VOID_RETURN;
+}
+
 static void network_init(void)
 {
 #ifdef HAVE_SYS_UN_H
@@ -2392,6 +2451,8 @@ static void network_init(void)
   int	arg;
 #endif
   DBUG_ENTER("network_init");
+
+  use_systemd_activated_sockets();
 
   if (MYSQL_CALLBACK_ELSE(thread_scheduler, init, (), 0))
     unireg_abort(1);			/* purecov: inspected */
@@ -2409,7 +2470,7 @@ static void network_init(void)
   if (!opt_disable_networking)
     DBUG_ASSERT(report_port != 0);
 #endif
-  if (!opt_disable_networking && !opt_bootstrap)
+  if (!opt_disable_networking && !opt_bootstrap && !systemd_sock_activation)
   {
     if (mysqld_port)
       activate_tcp_port(mysqld_port, &listen_sockets,
@@ -2423,7 +2484,7 @@ static void network_init(void)
   /*
   ** Create the UNIX socket
   */
-  if (mysqld_unix_port[0] && !opt_bootstrap)
+  if (mysqld_unix_port[0] && !opt_bootstrap && systemd_sock_activation==0)
   {
     MYSQL_SOCKET unix_sock= MYSQL_INVALID_SOCKET;
     size_t port_len;
@@ -5632,8 +5693,9 @@ int mysqld_main(int argc, char **argv)
 
   if (IS_SYSVAR_AUTOSIZE(&server_version_ptr))
     sql_print_information(ER_DEFAULT(ER_STARTUP), my_progname, server_version,
-                          (unix_sock_is_online ? mysqld_unix_port : (char*) ""),
-                          mysqld_port, MYSQL_COMPILATION_COMMENT);
+                          (systemd_sock_activation ? "Systemd socket activated ports" :
+                            (unix_sock_is_online ? mysqld_unix_port : (char*) "")),
+                          (systemd_sock_activation ? 0 : mysqld_port), MYSQL_COMPILATION_COMMENT);
   else
   {
     char real_server_version[2 * SERVER_VERSION_LENGTH + 10];
@@ -5644,8 +5706,9 @@ int mysqld_main(int argc, char **argv)
 
     sql_print_information(ER_DEFAULT(ER_STARTUP), my_progname,
                           real_server_version,
-                          (unix_sock_is_online ? mysqld_unix_port : (char*) ""),
-                          mysqld_port, MYSQL_COMPILATION_COMMENT);
+                          (systemd_sock_activation ? "Systemd socket activated ports" :
+                            (unix_sock_is_online ? mysqld_unix_port : (char*) "")),
+                          (systemd_sock_activation ? 0 : mysqld_port), MYSQL_COMPILATION_COMMENT);
   }
 
 #ifndef _WIN32
